requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
neo4j>=5.18
chromadb>=0.4.22
tree-sitter>=0.21
tree-sitter-language-pack>=0.6
python-dotenv>=1.0
openai>=1.30
anthropic>=0.25
google-generativeai>=0.7
tiktoken>=0.7
django-celery-beat>=2.6
django-celery-results>=2.5
psycopg2-binary>=2.9
gunicorn>=21.2
requests>=2.31
dj-database-url>=2.1

# Testing dependencies
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
